+++ database/migrations/2021_07_14_051258_add_router_status_to_device_configurations.php
@@ -26,9 +26,11 @@ class AddRouterStatusToDeviceConfigurations extends Migration
      */
     public function down()
     {
-        Schema::table('device_configurations', function (Blueprint $table) {
-            $table->dropColumn('router_status');
-            $table->dropColumn('router_update_time');
-        });
+        if (env('DB_CONNECTION', 'sqlite') !== 'sqlite') {
+            Schema::table('device_configurations', function (Blueprint $table) {
+                $table->dropColumn('router_status');
+                $table->dropColumn('router_update_time');
+            });
+        }
     }
 }
